:truck: Move history components into index view
